refactor: restructure project by layers (controller, service, repository) and move JWT claims to pkg
