+++ internal/controller/router.go
@@ -1,26 +1,26 @@
-package router
+package controller
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/kulikovroman08/reviewlink-backend/internal/auth/handler"
-	"github.com/kulikovroman08/reviewlink-backend/internal/auth/repository"
+	"github.com/kulikovroman08/reviewlink-backend/internal/controller/auth"
+	auth2 "github.com/kulikovroman08/reviewlink-backend/internal/repository/auth"
 )
 
-func SetupRouter(userRepo repository.UserRepository) *gin.Engine {
+func SetupRouter(userRepo auth2.UserRepository) *gin.Engine {
 	r := gin.Default()
 
 	r.GET("/health", func(c *gin.Context) {
 		c.JSON(200, gin.H{"status": "ok"})
 	})
 
-	authHandler := handler.NewHandler(userRepo)
+	authHandler := auth.NewHandler(userRepo)
 
 	// Auth endpoint
 	r.POST("/signup", authHandler.Signup)
 	r.POST("/login", authHandler.Login)
 
 	authorized := r.Group("/")
-	authorized.Use(handler.AuthMiddleware())
+	authorized.Use(auth.AuthMiddleware())
 	{
 		authorized.GET("/profiel", authHandler.GetProfiel)
 		authorized.POST("/reviews", authHandler.CreateReview)
