+++ internal/service/auth/model.go
