+++ pkg/jwt/claims.go
@@ -1,4 +1,4 @@
-package auth
+package jwt
 
 import "github.com/golang-jwt/jwt/v5"
